+++ pandasqt/__init__.py
@@ -3,7 +3,7 @@ __import__('pkg_resources').declare_namespace(__name__)
 __version__ = '0.1.2'
 
 #__all__ = ["DataFrameModel", "CustomDelegates", "DtypeComboDelegate"]
-from DataFrameModel import DataFrameModel
-from ColumnDtypeModel import DtypeComboDelegate
-from CustomDelegates import setDelegatesFromDtype, BigIntSpinboxDelegate, CustomDoubleSpinboxDelegate
-from DataSearch import DataSearch
+from pandasqt.models.DataFrameModel import DataFrameModel
+from pandasqt.models.ColumnDtypeModel import DtypeComboDelegate
+from pandasqt.views.CustomDelegates import setDelegatesFromDtype, BigIntSpinboxDelegate, CustomDoubleSpinboxDelegate
+from pandasqt.models.DataSearch import DataSearch
