+++ pandasqt/views/_ui/AddAttrDialog.py
