+++ pandasqt/views/_ui/AddAttributesDialog.ui
