+++ pandasqt/views/_ui/icons.qrc
