+++ pandasqt/views/_ui/icons/dialog-cancel.png
