+++ pandasqt/views/_ui/icons/dialog-ok-apply.png
