+++ pandasqt/views/_ui/icons/document-edit.png
