+++ pandasqt/views/_ui/icons/document-open.png
