+++ pandasqt/views/_ui/icons/document-save-as.png
