+++ pandasqt/views/_ui/icons/edit-table-delete-column.png
