+++ pandasqt/views/_ui/icons/edit-table-delete-row.png
