+++ pandasqt/views/_ui/icons/edit-table-insert-column-right.png
