+++ pandasqt/views/_ui/icons/edit-table-insert-row-below.png
