+++ pandasqt/views/_ui/icons/view-refresh.png
