+++ pandasqt/views/_ui/icons_rc.py
