+++ pandasqt/views/_ui/toolbar.qrc
